# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Community Donation Connect API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /requests:
    get:
      operationId: listRequests
      tags: [requests]
      summary: List receiver requests
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: urgency
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReceiverRequest"
    post:
      operationId: createRequest
      tags: [requests]
      summary: Create a receiver request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiverRequest"
  /requests/{id}:
    get:
      operationId: getRequest
      tags: [requests]
      summary: Get a request by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiverRequest"
    patch:
      operationId: updateRequest
      tags: [requests]
      summary: Update a receiver request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiverRequest"
    delete:
      operationId: deleteRequest
      tags: [requests]
      summary: Delete a receiver request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── MATCHING ENGINE ───────────────────────────────────────────────────
...
  /matches/find:
    post:
      operationId: findMatches
      tags: [matches]
      summary: Find matches for a donation or request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchSearch"
      responses:
        "200":
          description: Potential matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchResult"
  /matches/{id}/accept:
    patch:
      operationId: acceptMatch
      tags: [matches]
      summary: Accept a match
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Match accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
  /matches/{id}/confirm:
    patch:
      operationId: confirmMatch
      tags: [matches]
      summary: Confirm item received/transferred
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"

  # ── REVIEWS ───────────────────────────────────────────────────────────
  /reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/donor/{donorId}:
    get:
      operationId: getDonorReviews
      tags: [reviews]
      summary: Get reviews for a donor
      parameters:
        - name: donorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

  # ── NOTIFICATIONS ─────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    ReceiverRequest:
      type: object
      required: [id, receiverId, receiverName, category, itemName, quantity, urgency, status, radiusKm, createdAt]
      properties:
        id:
          type: integer
        receiverId:
          type: integer
        receiverName:
          type: string
        receiverVerified:
          type: boolean
        category:
          type: string
        itemName:
          type: string
        quantity:
          type: integer
        reason:
          type: ["string", "null"]
        urgency:
          type: string
          enum: [low, medium, high, critical]
        address:
          type: ["string", "null"]
        village:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        radiusKm:
          type: integer
        status:
          type: string
          enum: [active, matched, fulfilled, cancelled]
        distance:
          type: ["number", "null"]
        createdAt:
          type: string
    RequestInput:
      type: object
      required: [category, itemName, quantity, urgency, radiusKm]
      properties:
        category:
          type: string
        itemName:
          type: string
        quantity:
          type: integer
        reason:
          type: string
        urgency:
          type: string
          enum: [low, medium, high, critical]
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        radiusKm:
          type: integer
    RequestUpdate:
      type: object
      properties:
        itemName:
          type: string
        quantity:
          type: integer
        urgency:
          type: string
        radiusKm:
          type: integer
        status:
          type: string

    # ── MATCH SCHEMAS ─────────────────────────────────────────────────────
    Match:
      type: object
      required: [id, donationId, requestId, status, distanceKm, createdAt]
      properties:
        id:
          type: integer
        donationId:
          type: integer
        requestId:
          type: integer
        donorId:
          type: integer
        receiverId:
          type: integer
        donorName:
          type: string
        receiverName:
          type: string
        donorPhone:
          type: ["string", "null"]
        receiverPhone:
          type: ["string", "null"]
        donorEmail:
          type: ["string", "null"]
        receiverEmail:
          type: ["string", "null"]
        itemName:
          type: string
        category:
          type: string
        status:
          type: string
          enum: [pending, accepted, confirmed, completed, cancelled]
        distanceKm:
          type: number
        donorAccepted:
          type: boolean
        receiverAccepted:
          type: boolean
        deliveryType:
          type: ["string", "null"]
          enum: [pickup, delivery, "null"]
        createdAt:
          type: string
    MatchSearch:
      type: object
      required: [type, id]
      properties:
        type:
          type: string
          enum: [donation, request]
        id:
          type: integer
    MatchResult:
      type: object
      required: [id, category, itemName, distanceKm, urgency, trustScore, isVerified]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [donation, request]
        firstName:
          type: string
        village:
          type: ["string", "null"]
        category:
          type: string
        itemName:
          type: string
        distanceKm:
          type: number
        urgency:
          type: ["string", "null"]
        trustScore:
          type: integer
        isVerified:
          type: boolean

    # ── REVIEW SCHEMAS ────────────────────────────────────────────────────
    Review:
      type: object
      required: [id, matchId, donorId, receiverId, qualityRating, conditionRating, satisfactionRating, createdAt]
      properties:
        id:
          type: integer
        matchId:
          type: integer
        donorId:
          type: integer
        receiverId:
          type: integer
        qualityRating:
          type: integer
        conditionRating:
          type: integer
        satisfactionRating:
          type: integer
        feedback:
          type: ["string", "null"]
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [matchId, qualityRating, conditionRating, satisfactionRating]
      properties:
        matchId:
          type: integer
        qualityRating:
          type: integer
        conditionRating:
          type: integer
        satisfactionRating:
          type: integer
        feedback:
          type: string

    # ── NOTIFICATION SCHEMAS ──────────────────────────────────────────────
...